move file position and rename file
