+++ src/page-contents/index.ts
@@ -1,3 +1,3 @@
-import content2023 from "./2023/02_1";
+import content2023 from "./2023";
 
 export default [content2023];
